rename cluster enrichments
