+++ http/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/db/InstanceComponentSpec.scala
@@ -3,7 +3,7 @@ package org.broadinstitute.dsde.workbench.leonardo.db
 import org.broadinstitute.dsde.workbench.leonardo.{CommonTestData, GceInstanceStatus, GcsPathUtils, IP, RuntimeConfigId}
 import org.scalatest.FlatSpecLike
 import CommonTestData._
-import org.broadinstitute.dsde.workbench.leonardo.ClusterEnrichments.clusterEq
+import org.broadinstitute.dsde.workbench.leonardo.TestUtils.clusterEq
 
 import scala.concurrent.ExecutionContext.Implicits.global
