+++ http/src/test/scala/org/broadinstitute/dsde/workbench/leonardo/db/KubernetesClusterComponentSpec.scala
@@ -4,7 +4,7 @@ import java.time.Instant
 
 import org.broadinstitute.dsde.workbench.leonardo.KubernetesTestData._
 import org.broadinstitute.dsde.workbench.leonardo.CommonTestData._
-import org.broadinstitute.dsde.workbench.leonardo.ClusterEnrichments._
+import org.broadinstitute.dsde.workbench.leonardo.TestUtils._
 import org.broadinstitute.dsde.workbench.leonardo.{KubernetesClusterAsyncFields, KubernetesClusterStatus}
 import org.scalatest.FlatSpecLike
 
@@ -27,7 +27,6 @@ class KubernetesClusterComponentSpec extends FlatSpecLike with TestComponent {
 
     dbFutureValue(kubernetesClusterQuery.getActiveFullClusterByName(savedCluster1.googleProject, savedCluster1.clusterName)) shouldEqual  Some(savedCluster1)
     dbFutureValue(kubernetesClusterQuery.getActiveFullClusterByName(savedCluster2.googleProject, savedCluster2.clusterName)) shouldEqual Some(savedCluster2)
-
 
     //should delete the cluster and initial nodepool, hence '2' records deleted
     dbFutureValue(kubernetesClusterQuery.delete(savedCluster1.id)) shouldBe 2
